Add a go.mod file to support Go modules.
In order to work well with Go modules, packages should have a go.mod file and
semantic version tags on release versions.

Here I have provided a rudimentary go.mod file for the xz package. If and when
it is merged, the next desired step would be to tag and push a release
including this file.